+++ ActionUtilities macOS/Extensions/ColorExtensions.swift
@@ -230,7 +230,7 @@ extension NSColor {
         get {
             var (red, green, blue, alpha) = (CGFloat(0.0), CGFloat(0.0), CGFloat(0.0), CGFloat(0.0))
             self.getRed(&red, green: &green, blue: &blue, alpha: &alpha)
-            return (red: Int(red * 255), green: Int(green * 255), blue: Int(blue * 255), alpha: Int(alpha * 100))
+            return (red: Int(red * 255.0), green: Int(green * 255.0), blue: Int(blue * 255.0), alpha: Int(alpha * 100.0))
         }
     }
     
@@ -300,7 +300,7 @@ extension NSColor {
         get {
             var (hue, saturation, brightness, alpha) = (CGFloat(0.0), CGFloat(0.0), CGFloat(0.0), CGFloat(0.0))
             self.getRed(&hue, green: &saturation, blue: &brightness, alpha: &alpha)
-            return (hue: Int(hue * 360), saturation: Int(saturation * 100), brightness: Int(brightness * 100), alpha: Int(alpha * 100))
+            return (hue: Int(hue * 360.0), saturation: Int(saturation * 100.0), brightness: Int(brightness * 100.0), alpha: Int(alpha * 100.0))
         }
     }
     
@@ -356,7 +356,7 @@ extension NSColor {
         get {
             var (red, green, blue, alpha) = (CGFloat(0.0), CGFloat(0.0), CGFloat(0.0), CGFloat(0.0))
             self.getRed(&red, green: &green, blue: &blue, alpha: &alpha)
-            return (shade: Int(red * 255), alpha: Int(alpha * 100))
+            return (shade: Int(red * 255.0), alpha: Int(alpha * 100.0))
         }
     }
     
@@ -413,9 +413,9 @@ extension NSColor {
         }
         
         if alpha {
-            return prefix + String(format: "%02lX%02lX%02lX%02lX", lroundf(r * 255), lroundf(g * 255), lroundf(b * 255), lroundf(a * 255))
+            return prefix + String(format: "%02lX%02lX%02lX%02lX", lroundf(r * 255.0), lroundf(g * 255.0), lroundf(b * 255.0), lroundf(a * 255.0))
         } else {
-            return prefix + String(format: "%02lX%02lX%02lX", lroundf(r * 255), lroundf(g * 255), lroundf(b * 255))
+            return prefix + String(format: "%02lX%02lX%02lX", lroundf(r * 255.0), lroundf(g * 255.0), lroundf(b * 255.0))
         }
     }